test: Refactor & speed up
